+++ cadastro_usuario.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head>
+    <title>Cadastro</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/css/reset.css">
+    <link rel="stylesheet" href="/css/cadastro_menu/cadastro_menu.css">
+</head>
+    <body>
+        <h1 class="titulo">Registrar</h1>
+        <div class="container">
+            <form method="post" class="cadastro-usuario">
+                <label>Nome completo:</label>
+                <input type="text" class="cadastro-usuario__nome">
+                <label>Telefone</label>
+                <input type="number" name="telefone" class="cadastro-usuario__telefone">
+                <label>E-mail</label>
+                <input type="email" name="cadastro-usuario__email">
+                <label>Senha</label>
+                <input type="password" name="cadastro-usuario__senha">
+                <label>Confirme sua senha</label>
+                <input type="password" name="cadastro-usuario__confirmar-senha">
+                <button type="submit" class="cadastro-usuario__enviar">Enviar</button>
+            </form>
+        </div>
+    </body>
+</html>
