+++ css/login_menu/tela_login.css
@@ -1,27 +1,37 @@
 body {
     display: flex;
-    justify-content: center;
-    padding-top: 15%;
     background-color: #432534;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
 }
+.container { 
+    display: flex;
+    flex-direction: column; 
+    gap: 2rem;
+}
+
+.titulo {
+    text-align: center;
+    font-size: 2rem;
+    font-family: 'Outfit';
+    border-bottom: solid 2px;
+    color:  #0c041f;
+}
+
 .login-menu {
     display: flex;
     justify-content: center;
     align-items: center;
-    width: 40%;
     background-color: #efd6ac;
     padding: 4rem;
     border-radius: 10px;
 }
 
-.login-menu__form {
-    
-}
-
 .login-menu__label {
     border-radius: 3px;
     font-family: 'Outfit';
-    color:  #04151f;
+    color:  #000000;
 }
 
 .login-menu__input {
@@ -30,7 +40,7 @@ body {
 }
 
 .login-menu__botao {
-    background-color: #c44900;
+    background-color: #e95601;
     width: 100%;
     text-align: center;
     border-radius: 3px;
@@ -38,4 +48,15 @@ body {
 }
 .login-menu__botao:hover {
     cursor: pointer;
+    background-color: #7c2e00;
+}
+a {
+    display: block;
+    font-family: 'Outfit';
+    text-align: center;
+    margin: 0.3rem 0 0 0;
+}
+
+a:hover {
+    color: rgb(98, 0, 255);
 }
