+++ login.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="br">
+<html lang="pt-br">
     <head>
         <title>Login</title>
         <meta charset="UTF-8">
@@ -8,14 +8,19 @@
         <link rel="stylesheet" href="/css/login_menu/tela_login.css">
     </head>
     <body>
-        <!-- parte do login do usuário -->
-        <section class="login-menu">
-            <form action="http://localhost:3000/inicio" method="post" class="login-menu__form">
-                <label class="login-menu__label">Login: <input type="text" name="nome" class="login-menu__input"></label>
-                <label class="login-menu__label">Senha: <input type="password" name="senha" class="login-menu__input"></label>
-                <button type="submit" class="login-menu__botao">Entrar</button>
-            </form>
-        </section>
-        
+        <div class="container">
+            <h1 class="titulo">Gerenciamento de Estoque</h1>
+            <!-- parte do login do usuário -->
+            <section class="login-menu">
+                <form action="http://localhost:3000/inicio" method="post" class="login-menu__form">
+                    <label class="login-menu__label">Login:</label>
+                    <input type="text" name="nome" class="login-menu__input">
+                    <label class="login-menu__label">Senha:</label>
+                    <input type="password" name="senha" class="login-menu__input">
+                    <button type="submit" class="login-menu__botao">Entrar</button>
+                    <a href="http://localhost:3000/cadastro">Novo usuário?</a>
+                </form>
+            </section>
+        </div>
     </body>
 </html>
