+++ package.json
@@ -13,5 +13,8 @@
     "axios": "^1.9.0",
     "express": "^5.1.0",
     "http-server": "^14.1.1"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.10"
   }
 }
